+++ Source/equipment/buckler-devoted.js
@@ -1,11 +1,11 @@
 const EquipmentTemplate = require('../../Classes/EquipmentTemplate.js');
 const { addBlock, removeModifier, addModifier } = require('../combatantDAO.js');
 
-module.exports = new EquipmentTemplate("Buckler", "Grant an ally @{block} block and gain @{mod1Stacks} @{mod1}", "Block x@{critBonus}", "Earth", effect, ["Devoted Buckler", "Guarding Buckler", "Heavy Buckler"])
+module.exports = new EquipmentTemplate("Devoted Buckler", "Grant an ally @{block} block and @{mod1Stacks} @{mod1}", "Block x@{critBonus}", "Earth", effect, ["Guarding Buckler", "Heavy Buckler"])
 	.setCategory("Armor")
 	.setTargetingTags({ target: "single", team: "delver" })
 	.setModifiers([{ name: "Stagger", stacks: 1 }, { name: "Power Up", stacks: 25 }])
-	.setCost(200)
+	.setCost(350)
 	.setUses(10)
 	.setBlock(75);
 
@@ -18,6 +18,6 @@ function effect([target], user, isCrit, adventure) {
 		block *= critBonus;
 	}
 	addBlock(target, block);
-	addModifier(user, powerUp);
+	addModifier(target, powerUp);
 	return "";
 }