+++ Source/equipment/buckler-guarding.js
@@ -1,16 +1,16 @@
 const EquipmentTemplate = require('../../Classes/EquipmentTemplate.js');
-const { addBlock, removeModifier } = require('../combatantDAO.js');
+const { addBlock, removeModifier, addModifier } = require('../combatantDAO.js');
 
-module.exports = new EquipmentTemplate("Guarding Buckler", "Grant @{block} block to an ally and yourself", "Block x@{critBonus}", "Earth", effect, ["Heavy Buckler", "Urgent Buckler"])
+module.exports = new EquipmentTemplate("Guarding Buckler", "Grant @{block} block to an ally and yourself and gain @{mod1Stacks} @{mod1}", "Block x@{critBonus}", "Earth", effect, ["Devoted Buckler", "Heavy Buckler"])
 	.setCategory("Armor")
 	.setTargetingTags({ target: "single", team: "delver" })
-	.setModifiers([{ name: "Stagger", stacks: 1 }])
+	.setModifiers([{ name: "Stagger", stacks: 1 }, { name: "Power Up", stacks: 25 }])
 	.setCost(350)
 	.setUses(10)
 	.setBlock(75);
 
 function effect([target], user, isCrit, adventure) {
-	let { element, modifiers: [elementStagger], block, critBonus } = module.exports;
+	let { element, modifiers: [elementStagger, powerUp], block, critBonus } = module.exports;
 	if (user.element === element) {
 		removeModifier(user, elementStagger);
 		removeModifier(target, elementStagger);
@@ -20,5 +20,6 @@ function effect([target], user, isCrit, adventure) {
 	}
 	addBlock(target, block);
 	addBlock(user, block);
+	addModifier(user, powerUp);
 	return "";
 }