+++ Source/equipment/lance-base.js
@@ -1,28 +1,26 @@
 const EquipmentTemplate = require('../../Classes/EquipmentTemplate.js');
-const { dealDamage, addModifier, addBlock, getFullName } = require('../combatantDAO.js');
+const { dealDamage, addModifier, getFullName } = require('../combatantDAO.js');
 
-module.exports = new EquipmentTemplate("Guarding Sword", "Strike a foe for @{damage} @{element} damage, then gain @{block} block and @{mod1Stacks} @{mod1}", "Damage x@{critBonus}", "Earth", effect, ["Accelerating Sword", "Reckless Sword"])
+module.exports = new EquipmentTemplate("Lance", "Strike a foe for @{damage} @{element} damage and gain @{mod1Stacks} @{mod1}", "Damage x@{critBonus}", "Earth", effect, ["Guarding Lance", "Reckless Lance", "Slowing Lance"])
 	.setCategory("Weapon")
 	.setTargetingTags({ target: "single", team: "enemy" })
 	.setModifiers([{ name: "Stagger", stacks: 1 }, { name: "Power Up", stacks: 25 }])
-	.setCost(350)
+	.setCost(200)
 	.setUses(10)
-	.setDamage(75)
-	.setBlock(75);
+	.setDamage(75);
 
 function effect([target], user, isCrit, adventure) {
 	if (target.hp < 1) {
 		return ` ${getFullName(target, adventure.room.enemyTitles)} was already dead!`;
 	}
 
-	let { element, modifiers: [elementStagger, powerUp], damage, block, critBonus } = module.exports;
+	let { element, modifiers: [elementStagger, powerUp], damage, critBonus } = module.exports;
 	if (user.element === element) {
 		addModifier(target, elementStagger);
 	}
 	if (isCrit) {
 		damage *= critBonus;
 	}
-	addBlock(user, block);
 	addModifier(user, powerUp);
 	return dealDamage(target, user, damage, false, element, adventure);
 }