+++ Source/equipment/lance-slowing.js
@@ -1,11 +1,11 @@
 const EquipmentTemplate = require('../../Classes/EquipmentTemplate.js');
 const { dealDamage, addModifier, getFullName } = require('../combatantDAO.js');
 
-module.exports = new EquipmentTemplate("Sword", "Strike a foe for @{damage} @{element} damage (double increase from Power Up)", "Damage x@{critBonus}", "Earth", effect, ["Accelerating Sword", "Piercing Sword", "Vigilant Sword"])
+module.exports = new EquipmentTemplate("Slowing Lance", "Strike a foe for @{damage} @{element} damage, then inflict @{mod1Stacks} @{mod1} on them and gain @{mod2Stacks} @{mod2}", "Damage x@{critBonus}", "Earth", effect, ["Guarding Lance", "Reckless Lance"])
 	.setCategory("Weapon")
 	.setTargetingTags({ target: "single", team: "enemy" })
-	.setModifiers([{ name: "Stagger", stacks: 1 }])
-	.setCost(200)
+	.setModifiers([{ name: "Stagger", stacks: 1 }, { name: "Power Up", stacks: 25 }, { name: "Slow", stacks: 1 }])
+	.setCost(350)
 	.setUses(10)
 	.setDamage(75);
 
@@ -14,15 +14,14 @@ function effect([target], user, isCrit, adventure) {
 		return ` ${getFullName(target, adventure.room.enemyTitles)} was already dead!`;
 	}
 
-	let { element, modifiers: [elementStagger], damage, critBonus } = module.exports;
+	let { element, modifiers: [elementStagger, powerUp, slow], damage, critBonus } = module.exports;
 	if (user.element === element) {
 		addModifier(target, elementStagger);
 	}
-	const powerUpStacks = user.getModifierStacks("Power Up");
-	damage += powerUpStacks;
 	if (isCrit) {
 		damage *= critBonus;
-		damage += powerUpStacks;
 	}
+	addModifier(target, slow);
+	addModifier(user, powerUp);
 	return dealDamage(target, user, damage, false, element, adventure);
 }