+++ Source/equipment/sword-accelerating.js
@@ -1,10 +1,10 @@
 const EquipmentTemplate = require('../../Classes/EquipmentTemplate.js');
 const { dealDamage, addModifier, getFullName } = require('../combatantDAO.js');
 
-module.exports = new EquipmentTemplate("Accelerating Sword", "Strike a foe for @{damage} @{element} damage, then gain @{mod1Stacks} @{mod1} and @{mod2Stacks} @{mod2}", "Damage x@{critBonus}", "Earth", effect, ["Guarding Sword", "Reckless Sword"])
+module.exports = new EquipmentTemplate("Accelerating Sword", "Strike a foe for @{damage} @{element} damage (double increase from Power Up), then gain @{mod1Stacks} @{mod1}", "Damage x@{critBonus}", "Earth", effect, ["Piercing Sword", "Vigilant Sword"])
 	.setCategory("Weapon")
 	.setTargetingTags({ target: "single", team: "enemy" })
-	.setModifiers([{ name: "Stagger", stacks: 1 }, { name: "Power Up", stacks: 25 }, { name: "Quicken", stacks: 1 }])
+	.setModifiers([{ name: "Stagger", stacks: 1 }, { name: "Quicken", stacks: 1 }])
 	.setCost(350)
 	.setUses(10)
 	.setDamage(75);
@@ -14,14 +14,16 @@ function effect([target], user, isCrit, adventure) {
 		return ` ${getFullName(target, adventure.room.enemyTitles)} was already dead!`;
 	}
 
-	let { element, modifiers: [elementStagger, powerUp, quicken], damage, critBonus } = module.exports;
+	let { element, modifiers: [elementStagger, quicken], damage, critBonus } = module.exports;
 	if (user.element === element) {
 		addModifier(target, elementStagger);
 	}
+	const powerUpStacks = user.getModifierStacks("Power Up");
+	damage += powerUpStacks;
 	if (isCrit) {
 		damage *= critBonus;
+		damage += powerUpStacks;
 	}
 	addModifier(user, quicken);
-	addModifier(user, powerUp);
 	return dealDamage(target, user, damage, false, element, adventure);
 }