+++ Source/equipment/sword-piercing.js
@@ -1,11 +1,11 @@
 const EquipmentTemplate = require('../../Classes/EquipmentTemplate.js');
 const { dealDamage, addModifier, getFullName } = require('../combatantDAO.js');
 
-module.exports = new EquipmentTemplate("Sword", "Strike a foe for @{damage} @{element} damage (double increase from Power Up)", "Damage x@{critBonus}", "Earth", effect, ["Accelerating Sword", "Piercing Sword", "Vigilant Sword"])
+module.exports = new EquipmentTemplate("Piercing Sword", "Strike a foe for @{damage} @{element} unblockable damage (double increase from Power Up)", "Damage x@{critBonus}", "Earth", effect, ["Accelerating Sword", "Vigilant Sword"])
 	.setCategory("Weapon")
 	.setTargetingTags({ target: "single", team: "enemy" })
 	.setModifiers([{ name: "Stagger", stacks: 1 }])
-	.setCost(200)
+	.setCost(350)
 	.setUses(10)
 	.setDamage(75);
 
@@ -24,5 +24,5 @@ function effect([target], user, isCrit, adventure) {
 		damage *= critBonus;
 		damage += powerUpStacks;
 	}
-	return dealDamage(target, user, damage, false, element, adventure);
+	return dealDamage(target, user, damage, true, element, adventure);
 }