+++ Source/equipment/sword-vigilant.js
@@ -1,20 +1,21 @@
 const EquipmentTemplate = require('../../Classes/EquipmentTemplate.js');
-const { dealDamage, addModifier, getFullName } = require('../combatantDAO.js');
+const { dealDamage, addModifier, addBlock, getFullName } = require('../combatantDAO.js');
 
-module.exports = new EquipmentTemplate("Sword", "Strike a foe for @{damage} @{element} damage (double increase from Power Up)", "Damage x@{critBonus}", "Earth", effect, ["Accelerating Sword", "Piercing Sword", "Vigilant Sword"])
+module.exports = new EquipmentTemplate("Vigilant Sword", "Strike a foe for @{damage} @{element} damage (double increase from Power Up), then gain @{mod1Stacks} @{mod1}", "Damage x@{critBonus}", "Earth", effect, ["Accelerating Sword", "Piercing Sword"])
 	.setCategory("Weapon")
 	.setTargetingTags({ target: "single", team: "enemy" })
-	.setModifiers([{ name: "Stagger", stacks: 1 }])
-	.setCost(200)
+	.setModifiers([{ name: "Stagger", stacks: 1 }, { name: "Vigilance", stacks: 2 }])
+	.setCost(350)
 	.setUses(10)
-	.setDamage(75);
+	.setDamage(75)
+	.setBlock(75);
 
 function effect([target], user, isCrit, adventure) {
 	if (target.hp < 1) {
 		return ` ${getFullName(target, adventure.room.enemyTitles)} was already dead!`;
 	}
 
-	let { element, modifiers: [elementStagger], damage, critBonus } = module.exports;
+	let { element, modifiers: [elementStagger, vigilance], damage, block, critBonus } = module.exports;
 	if (user.element === element) {
 		addModifier(target, elementStagger);
 	}
@@ -24,5 +25,7 @@ function effect([target], user, isCrit, adventure) {
 		damage *= critBonus;
 		damage += powerUpStacks;
 	}
+	addBlock(user, block);
+	addModifier(user, vigilance);
 	return dealDamage(target, user, damage, false, element, adventure);
 }